:sparkles: Add coup de norminette
*
